refactored controllers

+++ controllers/user-controller.js
@@ -1,51 +1,30 @@
 const { User } = require("../models");
-
 const userController = {
   createUser({ body }, res) {
     User.create(body)
       .then((dbUserData) => res.json(dbUserData))
       .catch((err) => res.status(400).json(err));
   },
-
   getAllUser(req, res) {
     User.find({})
-      .populate({
-        path: "Thought",
-        select: "-__v",
-      })
-      .populate({
-        path: "friends",
-        select: "-__v",
-      })
+      .populate({ path: "Thought", select: "-__v" })
+      .populate({ path: "friends", select: "-__v" })
       .select("-__v")
-      .sort({
-        _id: -1,
-      })
+      .sort({ _id: -1 })
       .then((dbUserData) => res.json(dbUserData))
       .catch((err) => {
         console.log(err);
         res.status(500).json(err);
       });
   },
-
   getUserById({ params }, res) {
-    User.findOne({
-      _id: params.id,
-    })
-      .populate({
-        path: "Thought",
-        select: "-__v",
-      })
-      .populate({
-        path: "friends",
-        select: "-__v",
-      })
+    User.findOne({ _id: params.id })
+      .populate({ path: "Thought", select: "-__v" })
+      .populate({ path: "friends", select: "-__v" })
       .select("-__v")
       .then((dbUserData) => {
         if (!dbUserData) {
-          res.status(404).json({
-            message: "User not found",
-          });
+          res.status(404).json({ message: "User not found" });
           return;
         }
         res.json(dbUserData);
@@ -55,42 +34,26 @@ const userController = {
         res.status(400).json(err);
       });
   },
-
   updateUser({ params, body }, res) {
-    User.findOneAndUpdate(
-      {
-        _id: params.id,
-      },
-      body,
-      {
-        new: true,
-        runValidators: true,
-      }
-    )
+    User.findOneAndUpdate({ _id: params.id }, body, {
+      new: true,
+      runValidators: true,
+    })
       .then((dbUserData) => {
         if (!dbUserData) {
-          res.status(404).json({
-            message: "User not found",
-          });
+          res.status(404).json({ message: "User not found" });
           return;
         }
         res.json(dbUserData);
       })
       .catch((err) => res.json(err));
   },
-
   deleteUser({ params }, res) {
-    Thought.deleteMany({
-      userId: params.id,
-    })
+    Thought.deleteMany({ userId: params.id })
       .then(() => {
-        User.findOneAndDelete({
-          userId: params.id,
-        }).then((dbUserData) => {
+        User.findOneAndDelete({ userId: params.id }).then((dbUserData) => {
           if (!dbUserData) {
-            res.status(404).json({
-              message: "User not found",
-            });
+            res.status(404).json({ message: "User not found" });
             return;
           }
           res.json(dbUserData);
@@ -98,62 +61,34 @@ const userController = {
       })
       .catch((err) => res.json(err));
   },
-
   addFriend({ params }, res) {
     User.findOneAndUpdate(
-      {
-        _id: params.id,
-      },
-      {
-        $push: {
-          friends: params.friendId,
-        },
-      },
-      {
-        new: true,
-      }
+      { _id: params.id },
+      { $push: { friends: params.friendId } },
+      { new: true }
     )
-      .populate({
-        path: "friends",
-        select: "-__v",
-      })
+      .populate({ path: "friends", select: "-__v" })
       .select("-__v")
       .then((dbUserData) => {
         if (!dbUserData) {
-          res.status(404).json({
-            message: "User not found",
-          });
+          res.status(404).json({ message: "User not found" });
           return;
         }
         res.json(dbUserData);
       })
       .catch((err) => res.json(err));
   },
-
   deleteFriend({ params }, res) {
     User.findOneAndUpdate(
-      {
-        _id: params.id,
-      },
-      {
-        $pull: {
-          friends: params.friendId,
-        },
-      },
-      {
-        new: true,
-      }
+      { _id: params.id },
+      { $pull: { friends: params.friendId } },
+      { new: true }
     )
-      .populate({
-        path: "friends",
-        select: "-__v",
-      })
+      .populate({ path: "friends", select: "-__v" })
       .select("-__v")
       .then((dbUserData) => {
         if (!dbUserData) {
-          res.status(404).json({
-            message: "User not found",
-          });
+          res.status(404).json({ message: "User not found" });
           return;
         }
         res.json(dbUserData);
@@ -161,5 +96,4 @@ const userController = {
       .catch((err) => res.status(400).json(err));
   },
 };
-
 module.exports = userController;
